+++ .eslintrc.js
@@ -1,6 +1,11 @@
 module.exports = {
   extends: ['react-app'],
   rules: {
-    '@typescript-eslint/explicit-function-return-type': 'warn',
+    '@typescript-eslint/explicit-function-return-type': [
+      'warn',
+      {
+        allowTypedFunctionExpressions: true,
+      },
+    ],
   },
 };