+++ src/types.ts
@@ -1,3 +1,11 @@
 export type Type = 'info' | 'success' | 'fail' | 'loading';
 
 export type NoArgsReturnVoidFunction = () => void;
+
+export interface Message {
+  id: number;
+  type: Type;
+  content: string;
+  duration?: number;
+  onClose?: NoArgsReturnVoidFunction;
+}